Limit size of cib_t to 16 bits

+++ core/include/cib.h
@@ -21,6 +21,8 @@
 #ifndef __CIB_H
 #define __CIB_H
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -29,9 +31,9 @@ extern "C" {
  * @brief circular integer buffer structure
  */
 typedef struct {
-    unsigned int read_count;    /**< number of (successful) read accesses */
-    unsigned int write_count;   /**< number of (successful) write accesses */
-    unsigned int mask;          /**< Size of buffer -1, i.e. mask of the bits */
+    uint16_t read_count;    /**< number of (successful) read accesses */
+    uint16_t write_count;   /**< number of (successful) write accesses */
+    uint16_t mask;          /**< Size of buffer -1, i.e. mask of the bits */
 } cib_t;
 
 /**
@@ -44,9 +46,9 @@ typedef struct {
  *
  * @param[out] cib      Buffer to initialize.
  *                      Must not be NULL.
- * @param[in]  size     Size of the buffer, must not exceed MAXINT/2.
+ * @param[in]  size     Size of the buffer, must not exceed INT16_MAX/2.
  */
-static inline void cib_init(cib_t *__restrict cib, unsigned int size)
+static inline void cib_init(cib_t *__restrict cib, uint16_t size)
 {
     cib_t c = CIB_INIT(size);
     *cib = c;
